requirements.txt
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
requests==2.31.0
python-dateutil==2.8.2
newsapi-python==0.2.6
spacy==3.7.2
python-dotenv==1.0.0 